Add Guide for realtime lib
